moves code under src
