+++ package.json
@@ -4,7 +4,7 @@
   "description": "Slack API mocker",
   "main": "index.js",
   "scripts": {
-    "pretest": "standard index.js mocker/*.js lib/*.js test/*.js test/**/*.js examples/**/*.js --fix",
+    "pretest": "standard index.js src/*.js src/**/*.js test/*.js test/**/*.js examples/**/*.js --fix",
     "test": "istanbul cover --report lcov --dir ./coverage/ _mocha --recursive ./test/ --grep ./test/**/*.spec.js -- --colors --reporter spec",
     "examples": "mocha --recursive ./examples/test/ ",
     "ci": "npm test && cat ./coverage/lcov.info | coveralls"
